+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Color temperature control for native SteamVR headsets (e.g. Vive (Pro), Valve Index, Bigscreen Beyond)
 - Color temperature control to the current brightness automations.
 - Sunset/Sunrise based automations for brightness and color temperature.
+- Option for the shutdown sequence being triggered by being alone to only trigger while the sleep mode is enabled.
 
 ### Changed